[bashful] gist install crash was corrupting the registry file (#11)
Releases as v3.1.3

When `getGistVersion` returns an empty string, this corrupts the entire registry file. So half of the pkgman features don't work anymore.

This PR not only fixes why `getGistVersion` was breaking, it makes it so that it defaults to an empty string so that it does not corrupt the file.